CSP 5.9.0

The new changes are:
- Enable warnings as errors
- Remove unused serialization vector methods
- Only log to console if there is no log callback
- Fix SpaceTransform multiplication
- Emscripten release mode build warning
- Disable error warnings on Android builds
- Add Connected callback
- Let systems register callbacks before multiplayer connection is established
- Updated links to the docs to direct users towards the new URL
- Remove model component tests

+++ Library/src/Multiplayer/Script/ComponentBinding/TextSpaceComponentScriptInterface.cpp
@@ -40,8 +40,8 @@ DEFINE_SCRIPT_PROPERTY_VEC3(TextSpaceComponent, BackgroundColor);
 
 DEFINE_SCRIPT_PROPERTY_TYPE(TextSpaceComponent, bool, bool, IsBackgroundVisible);
 DEFINE_SCRIPT_PROPERTY_TYPE(TextSpaceComponent, BillboardMode, int64_t, BillboardMode);
-DEFINE_SCRIPT_PROPERTY_TYPE(TextSpaceComponent, uint32_t, uint32_t, Width);
-DEFINE_SCRIPT_PROPERTY_TYPE(TextSpaceComponent, uint32_t, uint32_t, Height);
+DEFINE_SCRIPT_PROPERTY_TYPE(TextSpaceComponent, float, uint32_t, Width);
+DEFINE_SCRIPT_PROPERTY_TYPE(TextSpaceComponent, float, uint32_t, Height);
 DEFINE_SCRIPT_PROPERTY_TYPE(TextSpaceComponent, bool, bool, IsVisible);
 DEFINE_SCRIPT_PROPERTY_TYPE(TextSpaceComponent, bool, bool, IsARVisible);
